+++ xis-plugin/src/main/java/one/xis/plugin/XISGenerateFrameworkComponentsTask.java
@@ -50,6 +50,7 @@ public class XISGenerateFrameworkComponentsTask extends DefaultTask {
                 "Modal",
                 "Router",
                 "Include",
+                "ChangeSet",
                 "Repository",
                 "MongoRepository"
         ));